Move constellation into its own crate.
